split tests into their own folder
